Added mixin to concentrate methods shared by content type / portlet / tile
